+++ CHANGELOG.md
@@ -1,4 +1,9 @@
-# Version 0.19.0 - 2020-??-??
+# Version 0.20.0 - 2020-??-??
+### Features
+- Add ability to enable / disable emotes globally
+- Add ability to manage emote packs with different state keys
+
+# Version 0.19.0 - 2020-09-21
 ### Features
 - Implemented ignore list
 - Jump to events in timeline: When tapping on a reply and when tapping a matrix.to link